requirements.txt
Flask == 1.1.1
gunicorn == 20.0.4
gcloud == 0.17.0
google-cloud-storage==1.28.1